+++ Example/NavigationExample/NavigationExample/Modular/ModularBoolView.swift
@@ -20,7 +20,7 @@ struct ModularBoolView: View {
     private var isMainShowing: Bool = false
 
     @State
-    var isTabShowing: Bool = false
+    private var isTabShowing: Bool = false
 
     @OptionalEnvironmentObject
     private var navigationStorage: NavigationStorage?
@@ -110,7 +110,7 @@ struct ModularBoolView: View {
             }
         }
         .padding()
-        .navigation(item: $stringForFirst) { stringValue in
+        .navigation(item: $stringForFirst, paramName: "firstString") { stringValue in
             Destination.first(stringValue)
         }
         .navigationAction(isActive: $isMainShowing) {