+++ Example/NavigationExample/NavigationExample/Standard/FirstView.swift
@@ -18,6 +18,11 @@ struct FirstView: View {
     @State
     private var isBoolShowed: Bool = false
 
+    @State
+    private var object: ObjectDTO? = nil
+    @State
+    private var objectDTO: ObjectDTO? = nil
+
     @Environment(\.isChange)
     private var isChange
 
@@ -40,6 +45,9 @@ struct FirstView: View {
                 Button("to Second with 22") {
                     numberForSecond = 22
                 }
+                Button("to Object") {
+                    object = .testValue1
+                }
                 Button("dismiss") {
                     presentationMode.wrappedValue.dismiss()
                 }
@@ -56,6 +64,29 @@ struct FirstView: View {
         }
         .navigationAction(item: $numberForSecond, paramName: "secondNumber", isRemovingParam: true) { numberValue in
             SecondView(number: numberValue, numberFromParent: $numberForSecond)
+        }
+        .navigationAction(item: $object, paramName: "object") { objectValue in
+            ObjectView(object: objectValue)
+        }
+        .navigation(item: $objectDTO, id: "object") { objectValue in
+            ObjectView(object: objectValue)
+        }
+        .navigateUrlParams("object") { path in
+            guard let id = path.getIntParam("object.id"),
+                  let name = path.getStringParam("object.name"),
+                  let rawDate = path.getStringParam("object.date"),
+                  let date = ObjectDTO.dateFormatter.date(from: rawDate)
+            else {
+                return
+            }
+            objectDTO = ObjectDTO(id: id, name: name, date: date)
+        } save: { path in
+            guard let object = objectDTO else {
+                return
+            }
+            path.pushIntParam("object.id", value: object.id)
+            path.pushStringParam("object.name", value: object.name)
+            path.pushStringParam("object.date", value: ObjectDTO.dateFormatter.string(from: object.date))
         }
     }
 }