+++ Sources/SUINavigation/Core/NavigationActionPath.swift
@@ -7,7 +7,9 @@
 
 import Foundation
 
-public typealias NavigateUrlParamsHandler = (_ path: NavigationActionPathProtocol) -> Void
+public typealias NavigateUrlParamsHandler = NavigateUrlParamsLoadHandler
+public typealias NavigateUrlParamsLoadHandler = (_ path: NavigationActionPathProtocol) -> Void
+public typealias NavigateUrlParamsSaveHandler = (_ path: NavigationActionSavePathProtocol) -> Void
 
 public protocol NavigationParameterValue {
     init?(_ description: String)
@@ -51,9 +53,25 @@ public protocol NavigationActionPathProtocol {
     func getParam<T: NavigationParameterValue>(_ name: String) -> T?
 }
 
-final public class NavigationActionPath: NavigationActionPathProtocol {
+public protocol NavigationActionSavePathProtocol {
+
+    func pushStringParam(_ name: String, value: String)
+    func replaceStringParam(_ name: String, value: String)
+
+    func pushIntParam(_ name: String, value: Int)
+    func replaceIntParam(_ name: String, value: Int)
+
+    func pushParam<T: NavigationParameterValue>(_ name: String, _ value: T)
+    func replaceParam<T: NavigationParameterValue>(_ name: String, _ value: T)
+}
+
+final public class NavigationActionPath {
     private(set) var path: [String]
     private var params: [NavigationParameter] = []
+
+    init(path: [String]) {
+        self.path = path
+    }
 
     init(url: String) {
         let pathEnd = url.firstIndex(of: "?") ?? url.endIndex
@@ -88,9 +106,32 @@ final public class NavigationActionPath: NavigationActionPathProtocol {
         }
     }
 
+    func pushParam(_ param: NavigationParameter) {
+        params.append(param)
+    }
+
+    func replaceParam(_ param: NavigationParameter) {
+        #warning("check the same...")
+        params.append(param)
+    }
+
+    func getURL() -> String {
+        let pathString = path.joined(separator: "/")
+        let paramsString = params.compactMap { param in
+            return "\(param.name)=\(param.value)"
+        }
+            .filter{ $0.isEmpty == false }
+            .joined(separator: "&")
+        if paramsString.isEmpty {
+            return pathString
+        } else {
+            return "\(pathString)?\(paramsString)"
+        }
+    }
+
 }
 
-extension NavigationActionPath {
+extension NavigationActionPath: NavigationActionPathProtocol {
 
     public func popStringParam(_ name: String) -> String? {
         var index = 0
@@ -140,5 +181,31 @@ extension NavigationActionPath {
             return nil
         }
         return T(param)
+    }
+}
+
+extension NavigationActionPath: NavigationActionSavePathProtocol {
+    public func pushStringParam(_ name: String, value: String) {
+        pushParam(NavigationParameter(name: name, value: value))
+    }
+
+    public func replaceStringParam(_ name: String, value: String) {
+        replaceParam(NavigationParameter(name: name, value: value))
+    }
+
+    public func pushIntParam(_ name: String, value: Int) {
+        pushParam(NavigationParameter(name: name, value: value.stringValue))
+    }
+
+    public func replaceIntParam(_ name: String, value: Int) {
+        replaceParam(NavigationParameter(name: name, value: value.stringValue))
+    }
+
+    public func pushParam<T>(_ name: String, _ value: T) where T : NavigationParameterValue {
+        pushParam(NavigationParameter(name: name, value: value.stringValue))
+    }
+
+    public func replaceParam<T>(_ name: String, _ value: T) where T : NavigationParameterValue {
+        replaceParam(NavigationParameter(name: name, value: value.stringValue))
     }
 }