+++ Sources/SUINavigation/Modifiers/NavigateUrlParamsModifier.swift
@@ -12,21 +12,35 @@ struct NavigateUrlParamsModifier: ViewModifier {
     var urlComponent: String
 
     var action: NavigateUrlParamsHandler
+    var save: NavigateUrlParamsSaveHandler?
 
     @OptionalEnvironmentObject
     var navigationStorage: NavigationStorage?
 
+    @State
+    private var uid: String? = nil
+
     func body(content: Content) -> some View {
         content
             .onAppear{
-                navigationStorage?.addChild(urlComponent, action)
+                guard uid == nil else {
+                    return
+                }
+                uid = navigationStorage?.addChild(urlComponent, action, save)
             }
+        if let uid {
+            let _ = navigationStorage?.updateChild(uid: uid, urlComponent, action, save)
+        }
     }
 }
 
 public extension View {
-    func navigateUrlParams(_ urlComponent: String, action: @escaping NavigateUrlParamsHandler) -> some View {
-        staticCheckUrlParams(urlComponent, action: action)
-        return navigationModifier(NavigateUrlParamsModifier(urlComponent: urlComponent, action: action))
+    func navigateUrlParams(
+        _ urlComponent: String,
+        action: @escaping NavigateUrlParamsHandler,
+        save: NavigateUrlParamsSaveHandler? = nil
+    ) -> some View {
+        staticCheckUrlParams(urlComponent, action: action, save: save)
+        return navigationModifier(NavigateUrlParamsModifier(urlComponent: urlComponent, action: action, save: save))
     }
 }